completed springboot upto exception handling

+++ src/main/java/com/ideas2it/employeemanagement/service/impl/ProjectServiceImpl.java
@@ -3,6 +3,7 @@ package com.ideas2it.employeemanagement.service.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 
@@ -10,11 +11,13 @@ import com.ideas2it.employeemanagement.model.ProjectDto;
 import com.ideas2it.employeemanagement.model.TechStackDto;
 import com.ideas2it.employeemanagement.repository.ProjectRepository;
 import com.ideas2it.employeemanagement.service.ProjectService;
+import com.ideas2it.employeemanagement.util.Constants;
+import com.ideas2it.employeemanagement.util.exception.EmployeeManagementException;
 import com.ideas2it.employeemanagement.util.mapper.Mapper;
 
 @Service
 public class ProjectServiceImpl implements ProjectService {
-	
+
 	@Autowired
 	private ProjectRepository projectRepository;
 
@@ -22,35 +25,49 @@ public class ProjectServiceImpl implements ProjectService {
 		return Mapper.convertIntoDto(projectRepository.save(Mapper.convertIntoEntity(projectDto)));
 	}
 
-	public List<ProjectDto> getProjects() {
-		return Mapper.convertIntoProjectsDto(projectRepository.findAll());
+	public List<ProjectDto> getProjects() throws EmployeeManagementException {
+		List<ProjectDto> projects = Mapper.convertIntoProjectsDto(projectRepository.findAll());
+		if (projects.isEmpty()) {
+			throw new EmployeeManagementException(Constants.NO_RECORD_FOUND, "404", HttpStatus.NOT_FOUND);
+		}
+		return projects;
 	}
 
-	public ProjectDto getProjectById(int id) {
-		return Mapper.convertIntoDto(projectRepository.findById(id).orElse(null));
+	public ProjectDto getProjectById(int id) throws EmployeeManagementException {
+		return Mapper.convertIntoDto(projectRepository.findById(id).orElseThrow(
+				() -> new EmployeeManagementException(Constants.PROJECT_NOT_FOUND, "404", HttpStatus.NOT_FOUND)));
 	}
 
-	public String deleteProjectById(int id) {
-		projectRepository.deleteById(id);
+	public String deleteProjectById(int id) throws EmployeeManagementException {
+		if (projectRepository.existsById(id)) {
+			projectRepository.deleteById(id);
+		} else {
+			throw new EmployeeManagementException(Constants.PROJECT_NOT_FOUND, "404", HttpStatus.NOT_FOUND);
+		}
 		return "deleted successfully " + id;
 	}
 
-	public String updateProject(ProjectDto projectDto, int id) {
+	public String updateProject(ProjectDto projectDto, int id) throws EmployeeManagementException {
 		String message = null;
 		if (projectRepository.existsById(id)) {
-			ProjectDto existingProject = Mapper.convertIntoDto(projectRepository.findById(id).orElse(null));
-			existingProject.setName(projectDto.getName());
-			existingProject.setStartDate(projectDto.getStartDate());
-			existingProject.setEndDate(projectDto.getEndDate());
-			if (null != projectDto.getTechStacks()) {
-				List<TechStackDto> input = existingProject.getTechStacks();
-				input.addAll(projectDto.getTechStacks());
-				existingProject.setTechStacks(input);
+			ProjectDto existingProject = getProjectById(id);
+			if (null != existingProject) {
+				existingProject.setName(projectDto.getName());
+				existingProject.setStartDate(projectDto.getStartDate());
+				existingProject.setEndDate(projectDto.getEndDate());
+				if (null != projectDto.getTechStacks()) {
+					List<TechStackDto> input = existingProject.getTechStacks();
+					input.addAll(projectDto.getTechStacks());
+					existingProject.setTechStacks(input);
+				}
+				if (existingProject.equals(projectRepository.save(Mapper.convertIntoEntity(projectDto)))) {
+					message = projectDto.getName() + " Update Successfully";
+				} else {
+					message = projectDto.getName() + " Update UnSuccessfull";
+				}
 			}
-			projectRepository.save(Mapper.convertIntoEntity(existingProject));	
-		    message = projectDto.getName() + " Update Successfully";
 		} else {
-			message = "Project not found";
+			throw new EmployeeManagementException(Constants.PROJECT_NOT_FOUND, "404", HttpStatus.NOT_FOUND);
 		}
 		return message;
 	}
