+++ src/main/java/com/ideas2it/employeemanagement/service/impl/TechStackServiceImpl.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -12,12 +13,13 @@ import com.ideas2it.employeemanagement.model.TechStack;
 import com.ideas2it.employeemanagement.model.TechStackDto;
 import com.ideas2it.employeemanagement.repository.TechStackRepository;
 import com.ideas2it.employeemanagement.service.TechStackService;
+import com.ideas2it.employeemanagement.util.Constants;
 import com.ideas2it.employeemanagement.util.exception.EmployeeManagementException;
 import com.ideas2it.employeemanagement.util.mapper.Mapper;
 
 @Service
 public class TechStackServiceImpl implements TechStackService {
-	
+
 	@Autowired
 	private TechStackRepository techStackRepository;
 
@@ -27,35 +29,43 @@ public class TechStackServiceImpl implements TechStackService {
 
 	public List<TechStackDto> getTechStacks() throws EmployeeManagementException {
 		List<TechStackDto> techStacks = Mapper.convertIntoTechStacksDto(techStackRepository.findAll());
-		if(!techStacks.isEmpty()) {
-			return techStacks;
-		} else {
-			throw new EmployeeManagementException("No Records Found");
+		if (techStacks.isEmpty()) {
+			throw new EmployeeManagementException(Constants.NO_RECORD_FOUND, "404", HttpStatus.NOT_FOUND);
 		}
+		return techStacks;
 	}
 
-	public TechStackDto getTechStackById(int id) {
-		return Mapper.convertIntoDto(techStackRepository.findById(id).orElse(null));
+	public TechStackDto getTechStackById(int id) throws EmployeeManagementException {
+		return Mapper.convertIntoDto(techStackRepository.findById(id).orElseThrow(
+				() -> new EmployeeManagementException(Constants.TECH_STACK_NOT_FOUND, "404", HttpStatus.NOT_FOUND)));
 	}
 
-	public String deleteTechStackById(int id) {
-		techStackRepository.deleteById(id);
+	public String deleteTechStackById(int id) throws EmployeeManagementException {
+		if (techStackRepository.existsById(id)) {
+			techStackRepository.deleteById(id);
+		} else {
+			throw new EmployeeManagementException(Constants.TECH_STACK_NOT_FOUND, "404", HttpStatus.NOT_FOUND);
+		}
 		return "deleted successfully " + id;
 	}
 
-	public String updateTechStack(TechStackDto techStackDto, int id) {
+	public String updateTechStack(TechStackDto techStackDto, int id) throws EmployeeManagementException {
 		String message = null;
 		if (techStackRepository.existsById(id)) {
-			TechStackDto existingTechStack = Mapper.convertIntoDto(techStackRepository.findById(id).orElse(null));
-			existingTechStack.setName(techStackDto.getName());
-			existingTechStack.setVersion(techStackDto.getVersion());
-			techStackRepository.save(Mapper.convertIntoEntity(existingTechStack));	
-		    message = techStackDto.getName() + " Update Successfully";
+			TechStackDto existingTechStack = getTechStackById(id);
+			if (null != existingTechStack) {
+				existingTechStack.setName(techStackDto.getName());
+				existingTechStack.setVersion(techStackDto.getVersion());
+				if (existingTechStack.equals(techStackRepository.save(Mapper.convertIntoEntity(techStackDto)))) {
+					message = techStackDto.getName() + " Update Successfully";
+				} else {
+					message = techStackDto.getName() + " Update UnSuccessfull";
+				}
+			}
 		} else {
-			message = "TechStack not found";
+			throw new EmployeeManagementException(Constants.TECH_STACK_NOT_FOUND, "404", HttpStatus.NOT_FOUND);
 		}
 		return message;
 	}
 
 }
-
